! Import Cutting Stock project. This is a project for Mr. Tuan's company.
The cutting stock problem is an optimization problem, or more specifically, an integer linear programming problem. It arises from many applications in industry. Imagine that you work in a paper mill and you have a number of rolls of paper of fixed width waiting to be cut, yet different customers want different numbers of rolls of various-sized widths. How are you going to cut the rolls so that you minimize the waste (amount of left-overs)?

Solving this problem to optimality can be economically significant: a difference of 1% for a modern paper machine can be worth more than US$ 1 million per annum.